Final Push - Ready for Review
Changed file names and added author name and date.   Added JS that allows for each food item to be added to the cart, along with its description, and will say the number of orders. Should the food item have been added already, it will instead increase the number of orders. Added comments.